refactor(gateway): move Brave search into its own submodule
Pure movement. No behavior change: `search_brave` and `parse_brave_hits` are
byte-identical apart from `pub(super)` on the first, and the test count is
unchanged at 77.

`fetcher.rs` was 712 lines and the 429 work that follows pushes it past the
repo's 800-line limit. Everything Brave-shaped is one coherent unit — endpoint,
request, response parsing — so it becomes `fetcher/brave.rs`; `fetcher.rs` keeps
the keyless DuckDuckGo tier, the SSRF screen, and `fetch`.

Two symbols sat inside the moved range without belonging to it and stay behind:
`fetch_err_brief`, which formats any `FetchError` for the fallback log, and the
`search_tier1` doc comment. Both were caught by listing which symbols each file
defines rather than by trusting the text boundaries the move was cut on.
